Change variable type for lower and upper bounds

+++ solutionmethod.h
@@ -9,7 +9,7 @@ typedef unsigned int uint;
 class SolutionMethod
 {
 public:
-    SolutionMethod(Expression *e1, Expression *e2, QtCharts::QChart *chart, int lowerBound, int upperBound, uint steps);
+    SolutionMethod(Expression *e1, Expression *e2, QtCharts::QChart *chart, double lowerBound, double upperBound, uint steps);
     virtual void nextStep() = 0;
     virtual void previousStep() = 0;
     uint getCurrentStep();
@@ -19,9 +19,9 @@ protected:
     QtCharts::QChart *chart;
     uint currentStep;
     Expression *e1, *e2;
-    int lowerBound;
-    int upperBound;
-     bool isIncreasingFunction;// could have problem in periodic functions if lowerBound and upperBound are too far
+    double lowerBound;
+    double upperBound;
+    bool isIncreasingFunction;// could have problem in periodic functions if lowerBound and upperBound are too far
 private:
     uint totalSteps;
 };
